Add TserverLoggingTest and make configuration changes
Add TserverLogging test to verify that queryLogIterator is created and
added to the iterator stack when desired. Additionally, move
tserver.logging.active property to QueryParameters, and ensure that if
specified, the query parameter will override the configuration settings.
Additionally, make it possible to set the tserver active logging flag
from a ShardQueryLogic configuration setting in the
QueryLogicFactory.xml file.